refactor: queryForObject 와 query 중복로직 추출
