+++ app/src/test/java/com/techcourse/service/MockUserHistoryDao.java
@@ -2,7 +2,7 @@ package com.techcourse.service;
 
 import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.UserHistory;
-import nextstep.jdbc.DataAccessException;
+import nextstep.jdbc.exception.DataAccessException;
 import nextstep.jdbc.JdbcTemplate;
 
 public class MockUserHistoryDao extends UserHistoryDao {
