--- jdbc/src/main/java/nextstep/jdbc/DataAccessException.java
@@ -1,26 +1,0 @@
-package nextstep.jdbc;
-
-public class DataAccessException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    public DataAccessException() {
-        super();
-    }
-
-    public DataAccessException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-
-    public DataAccessException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DataAccessException(String message) {
-        super(message);
-    }
-
-    public DataAccessException(Throwable cause) {
-        super(cause);
-    }
-}
