+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -2,9 +2,11 @@ package nextstep.jdbc;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
+import nextstep.jdbc.utils.DataAccessUtils;
 
 public class JdbcTemplate extends JdbcSupporter{
 
@@ -17,24 +19,22 @@ public class JdbcTemplate extends JdbcSupporter{
     }
 
     public <T> T queryForObject(final String sql, RowMapper<T> rowMapper, final Object...args) {
-        return execute(sql, preparedStatement -> {
-            final ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-                return rowMapper.mapRow(resultSet, resultSet.getRow());
-            }
-            return null;
-        }, args);
+        final List<T> results = execute(sql, preparedStatement -> getResult(rowMapper, preparedStatement), args);
+        return DataAccessUtils.nullableSingleResult(results);
     }
 
     public <T> List<T> query(final String sql, final RowMapper<T> rowMapper, final Object... args) {
-        return execute(sql, preparedStatement -> {
-            final ResultSet resultSet = preparedStatement.executeQuery();
-            int row = 0;
-            final List<T> results = new ArrayList<>();
-            while (resultSet.next()) {
-                results.add(rowMapper.mapRow(resultSet, row++));
-            }
-            return results;
-        }, args);
+        return execute(sql, preparedStatement -> getResult(rowMapper, preparedStatement), args);
+    }
+
+    private <T> List<T> getResult(final RowMapper<T> rowMapper,
+                                  final PreparedStatement preparedStatement) throws SQLException {
+        final ResultSet resultSet = preparedStatement.executeQuery();
+        int row = 0;
+        final List<T> results = new ArrayList<>();
+        while (resultSet.next()) {
+            results.add(rowMapper.mapRow(resultSet, row++));
+        }
+        return results;
     }
 }
