+++ jdbc/src/main/java/nextstep/jdbc/exception/DataAccessException.java
@@ -1,0 +1,10 @@
+package nextstep.jdbc.exception;
+
+public class DataAccessException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public DataAccessException() {
+        super();
+    }
+}
