+++ jdbc/src/main/java/nextstep/jdbc/exception/EmptyResultDataAccessException.java
@@ -1,0 +1,7 @@
+package nextstep.jdbc.exception;
+
+public class EmptyResultDataAccessException extends RuntimeException{
+    public EmptyResultDataAccessException(final String message) {
+        super(message);
+    }
+}
