+++ jdbc/src/main/java/nextstep/jdbc/exception/IncorrectResultSizeDataAccessException.java
@@ -1,0 +1,7 @@
+package nextstep.jdbc.exception;
+
+public class IncorrectResultSizeDataAccessException extends RuntimeException {
+    public IncorrectResultSizeDataAccessException(final String message) {
+        super(message);
+    }
+}
